% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appeears_wrapper.R
\name{appeears_fetch_bundle}
\alias{appeears_fetch_bundle}
\title{Fetch Bundle}
\usage{
appeears_fetch_bundle(task_id, token)
}
\arguments{
\item{task_id}{A task id as per the response given when executing the appeears_start_task function. Expects
a list which includes a variable, 'task_id'.}

\item{token}{An authentication token associated with the submitted task, as per the appeears_start_session function.
Expects a list with a a variable, 'token'}
}
\value{
A list representing the AppEEARS server response, which should include a set of bundle information, including
all files/file ids associated with the bundle.
}
\description{
Returns the bundle associated with a specified task.
}
