% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appeears_wrapper.R
\name{appeears_start_session}
\alias{appeears_start_session}
\title{Start AppEEARS Session}
\usage{
appeears_start_session(username, password)
}
\arguments{
\item{username}{Username with which to login}

\item{password}{Password with which to login}
}
\value{
Returns a list representing the response from the AppEEARS server. Includes a variable 'token', necessary
for making further requests
}
\description{
Provides an authentication token given a username/password valid with the NASA EarthDATA system.
Authentication token can be used to make further calls to the AppEEARS server.
}
