% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appeears_wrapper.R
\name{appeears_start_task}
\alias{appeears_start_task}
\title{Start Task}
\usage{
appeears_start_task(token, task_name, start_date, end_date, product, layers,
  type, points)
}
\arguments{
\item{token}{An authentication token associated with the submitted task, as per the appeears_start_session function.
Expects a list with a a variable, 'token'}

\item{task_name}{User specified name to give the task}

\item{start_date}{Used in conjunction with end_date to specify a date range for the search. Should be a string in MM-DD-YYYY
format}

\item{end_date}{Used in conjunction with start_date to specify a date range for the search. Should be a string in MM-DD-YYYY
format}

\item{product}{Name of the data product to specify in the request
A list of available products is available here: https://lpdaacsvc.cr.usgs.gov/appeears/products
Takes a string in the format product_name.version_number}

\item{layers}{layer names to pull from the product as specified in the product parameter. A list of layers
available with each product can be found here: https://lpdaacsvc.cr.usgs.gov/appeears/products
Takes a vecotor of strings, all of which should perfectly match the layer names specified in the above URL.}
}
\value{
Returns a list representing the AppEEARS server response. On a valid request, includes a variable
'task_id' which can be used in further requests to see the status of the task or to retrieve associated bundle.
}
\description{
Executes a task on the AppEEARS server for later retrieval.
Currently this only accomodates area search with a fixed set of points for the vector. Also, although multiple
layers can be specified they must all belong to the same product. So there is room for this to grow.
}
