% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appeears_wrapper.R
\name{appeears_task_is_done}
\alias{appeears_task_is_done}
\title{Is Task Done}
\usage{
appeears_task_is_done(token, task_id)
}
\arguments{
\item{token}{An authentication token associated with the submitted task, as per the appeears_start_session function.
Expects a list with a a variable, 'token'}

\item{task_id}{A task id as per the response given when executing the appeears_start_task function. Expects
a list which includes a variable, 'task_id'.}
}
\value{
A boolean representing whether or not the task is complete.
}
\description{
Returns true/false whether or not a given task is done and ready for retrieval.
}
