% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appeears_wrapper.R
\name{appeears_task_status}
\alias{appeears_task_status}
\title{Get Task Status}
\usage{
appeears_task_status(token, task_id)
}
\arguments{
\item{token}{An authentication token associated with the submitted task, as per the appeears_start_session function.
Expects a list with a a variable, 'token'}

\item{task_id}{A task id as per the response given when executing the appeears_start_task function. Expects
a list which includes a variable, 'task_id'.}
}
\value{
A list representing the AppEEARS server response which should include the 'percent complete' status of the
task in the task processing workflow.
}
\description{
Returns information about the status of a particular task. This allows a script to make decisions about what to do
given the state of on ongoing task but doesn't provide information about whether or not the task is complete. See
'Is Task Done' for info on that.
}
